[router-advertisement] fix implicit conversion loss (#6559)

+++ src/core/border_router/router_advertisement.hpp
@@ -112,7 +112,7 @@ public:
      * @param[in]  aSize  The size of the option in unit of 1 byte.
      *
      */
-    void SetSize(uint16_t aSize) { mLength = (aSize + kLengthUnit - 1) / kLengthUnit; }
+    void SetSize(uint16_t aSize) { mLength = static_cast<uint8_t>((aSize + kLengthUnit - 1) / kLengthUnit); }
 
     /**
      * This method returns the size of the option (in bytes).
